Sig-6: Created new settings.dart file and added settings class
